Add pi3-miniuart-bt to config.txt
This device tree overlay swaps the assignment of the two UARTs. With this
change the primary UART ("dev/ttyAMA0") is assigned to the GPIO pins, the
mini UART ("dev/ttyS0") is assigned to the onboard bluetooth chip.

For more information on the RPi UARTs see https://www.raspberrypi.org/documentation/configuration/uart.md